+++ src/main/java/com/healthCare/healthCareDataBase/Repository/MedicamentStockRepository.java
@@ -8,22 +8,22 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.healthCare.healthCareDataBase.Model.MedicamentStock;
 
-public interface MedicamentStockRepository extends JpaRepository <MedicamentStock,Integer>{
+public interface MedicamentStockRepository extends JpaRepository <MedicamentStock,Long>{
 
 	@Modifying
 	@Transactional
 	@Query(value="INSERT INTO medicament_stocks (medicament_id,medicament_stock_qte, pharmacy_id,medicament_stock_add_date) VALUES (?1,?2,?3,?4)",nativeQuery=true)
-	void addMedicamentStock(Integer medicamentId, Integer medicamentStockQte, Integer pharmacyId, String format);
+	void addMedicamentStock(Long medicamentId, Long medicamentStockQte, Long pharmacyId, String format);
 
 	@Query(value="select m.medicament_id from medicament_stocks m where m.medicament_id=?1 and m.pharmacy_id=?2",nativeQuery=true)
-	Integer checkIfStockExist(Integer medicamentId, Integer pharmacyId);
+	Long checkIfStockExist(Long medicamentId, Long pharmacyId);
 	
 	@Query(value="select m.medicament_stock_qte from medicament_stocks m where m.medicament_id=?1 and m.pharmacy_id=?2",nativeQuery=true)
-	Integer getStockQte(Integer medicamentId, Integer pharmacyId);
+	Long getStockQte(Long medicamentId, Long pharmacyId);
 	
 	@Modifying
 	@Transactional
 	@Query(value="update medicament_stocks ms set ms.medicament_stock_qte=?2 where ms.medicament_id=?1 and ms.pharmacy_id=?3",nativeQuery=true)
-	void updateMedicamentStock(Integer medicamentId, Integer medicamentStockQte, Integer pharmacyId);
+	void updateMedicamentStock(Long medicamentId, Long medicamentStockQte, Long pharmacyId);
 
 }
