+++ src/main/java/com/healthCare/healthCareDataBase/Repository/ValidationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.healthCare.healthCareDataBase.Model.Validation;
 
-public interface ValidationRepository extends JpaRepository<Validation, Integer>{
+public interface ValidationRepository extends JpaRepository<Validation, Long>{
 
 }
